add back tests for vari long double

+++ test/unit/math/rev/core/vari_test.cpp
@@ -10,7 +10,7 @@ TEST(AgradRev, insertion_operator) {
 }
 
 TEST(AgradRev, long_double_test) {
-  stan::math::vari_value<float> v(5);
+  stan::math::vari_value<long double> v(5);
   std::stringstream ss;
   ss << &v;
   EXPECT_EQ("5:0", ss.str());
